Support/test ansible 1.3-1.9.
This isn't testing every single version of ansible, but the latest
version in the 1.3.x series through the 1.9.x series. It's tedious
enough supporting this range with tox.

Older versions of ansible (1.7 and older, i.e. prior to 1.8) don't allow
there to be a .py extension on the module name, and making a symlink
fixed that.

Other older versions of ansible (1.4 and older, i.e. prior to 1.5) were
having trouble with py.test's IO capturing, so that's been turned off by
default. If there's ever a reason to turn it back on, it can get turned
back on.

Ansible 1.3 is the oldest version supported by pytest-ansible, so that's
that arbitrary line drawn.