Check idempotency by marking tests as idempotent.
An idempotent test means that you can run the runit_sv module twice and
there will be changes the first time but not the second time. Not all
tests are idempotent because some check for failures. This is much
easier than having to write separate idempotency guarantee tests because
they would pretty much all be duplicates of the existing tests.